chore: restructure repository to use src-layout and move tests to repo root
