+++ src/quickwrench_api/wsgi.py
@@ -4,4 +4,5 @@ from django.core.wsgi import get_wsgi_application
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "quickwrench_api.settings")
 
+
 application = get_wsgi_application()
